DEV: Update pluginConsequence module to build from a docker image
Provided runtimes were updated, deprecating the old version of this
module. To avoid this in future, a docker container has been created
that builds the equivalent as a docker image and uses that to deploy
the lambda function.

* Added docker_images.tf and docker_prep.py - defines the ecr and
computes hash on files to determine any changes to the image

* Updated lambda function in main.tf - No longer using fork of
terraform-aws-lambda, though this will likely be changed once
support is added

* Updated sns.tf to reflect attributes from non-forked lambda source

* Added variables.tf with region as it is needed to define ecr

* Updated versions.tf - made aws version explicit and added docker

* Added Dockerfile - configures required environment for VEP

* Added bootstrap.sh - handles lambda function initiation and passing
of sns notifications to VEP

* Updated VEP.pm handler to support new sns notification passing

* Updated VEP.pm handler to reference location awscli is installed to
in docker container

* Updated VEP.pm to only write to .tsv when the results array contains
elements

+++ lambda/pluginConsequence/docker/VEP.pm
@@ -47,6 +47,7 @@ Bio::EnsEMBL::Variation::Utils::VEP - Methods used by the Variant Effect Predict
 
 =cut
 
+package VEP;
 
 use strict;
 use warnings;
@@ -80,16 +81,10 @@ my $mirnaFile =  $ENV{'MIRNA_REFERENCE'};
 my $outputLocation =  $ENV{'SVEP_REGIONS'};
 my $tempLocation =  $ENV{'SVEP_TEMP'};
 sub handle {
-    my ($payload, $context) = @_;
-    #print Dumper $payload;
-    #my @data = decode_json($payload->{'Message'}); # This line is for individual lambda testing
-    #my $data = decode_json($payload->{'Message'});
-    #event['Records'][0]['Sns']['Message']
-    my $event = $payload->{'Records'}[0];
-    my $sns = $event->{'Sns'};
+    my ($payload) = @_;
+    my $event = decode_json($payload);
+    my $sns = $event->{Records}[0]{Sns};
     ##########################################update
-    #my @data = decode_json($sns->{'Message'});
-    #my $id = $sns->{'MessageId'};
     my $message = decode_json($sns->{'Message'}); #might have to remove decode_json
     my @data = $message->{'snsData'};
     my $tempFileName = $message->{'tempFileName'};
@@ -100,8 +95,8 @@ sub handle {
     my $chr = $data[0][0]->{'chrom'};
     #print($chr);
     my $fasta ='Homo_sapiens.GRCh38.dna.chromosome.'.$chr.'.fa.gz';
-    system("/opt/awscli/aws s3 cp $fastaLocation /tmp/ --recursive  --exclude '*'  --include '$fasta*'");
-    system("/opt/awscli/aws s3 cp $fastaLocation /tmp/ --recursive  --exclude '*'  --include '$spliceFile*'");
+    system("/usr/bin/aws s3 cp $fastaLocation /tmp/ --recursive  --exclude '*'  --include '$fasta*'");
+    system("/usr/bin/aws s3 cp $fastaLocation /tmp/ --recursive  --exclude '*'  --include '$spliceFile*'");
     my @results;
     while(@data){
       my $region = shift @data;
@@ -115,20 +110,24 @@ sub handle {
         }
       }
     }
-    #my $filename = "/tmp/test.tsv";
-    my $filename = "/tmp/".$tempFileName.".tsv";
-    open(my $fh, '>', $filename) or die "Could not open file '$filename' $!";
-    print $fh join("\n", @results);
-    close $fh;
+    if(scalar @results) {
+      #my $filename = "/tmp/test.tsv";
+      my $filename = "/tmp/".$tempFileName.".tsv";
+      open(my $fh, '>', $filename) or die "Could not open file '$filename' $!";
+      print $fh join("\n", @results);
+      close $fh;
 
-    my $out = 's3://'.$outputLocation.'/';
-    system("/opt/awscli/aws s3 cp $filename $out");
-    #print("Done Copying");
-    my $tempOut = 's3://'.$tempLocation.'/'.$tempFileName;
-    system("/opt/awscli/aws s3 rm $tempOut");
-    print("cleaning tmp");
-    system("rm -r /tmp/*"); #cleaning tmp after each use
-    print("Done Copying");
+      my $out = 's3://'.$outputLocation.'/';
+      system("/usr/bin/aws s3 cp $filename $out");
+      #print("Done Copying");
+      my $tempOut = 's3://'.$tempLocation.'/'.$tempFileName;
+      system("/usr/bin/aws s3 rm $tempOut");
+      print("cleaning tmp");
+      system("rm -r /tmp/*"); #cleaning tmp after each use
+      print("Done Copying");
+    } else {
+      print("Nothing to copy")
+    }
 }
 
 # parse a line of VCF input into a variation feature object
@@ -464,7 +463,7 @@ sub parse_vcf {
         my $mirnaLocalFile = "/tmp/".$mirnaFile;
         unless (-e $mirnaLocalFile) {
           print "File Doesn't Exist in local directory!";
-          system("/opt/awscli/aws s3 cp $fastaLocation /tmp/ --recursive  --exclude '*'  --include '$mirnaFile*'");
+          system("/usr/bin/aws s3 cp $fastaLocation /tmp/ --recursive  --exclude '*'  --include '$mirnaFile*'");
         }
         if($rows[1] eq "mirbase"){
           my $file = "sorted_filtered_mirna.gff3.gz";
@@ -567,7 +566,6 @@ sub parse_vcf {
         $cons = '5_prime_UTR_variant,'.$cons;
         }
       }
-
       my $line = $rank."\t".('.')."\t".$chr.':'.$start.'-'.$end."\t".$alt."\t".$cons."\t".$info{gene_name}."\t".$info{gene_id}."\t".$rows[2]."\t".
       $info{transcript_id}.".".$info{transcript_version}."\t".$info{transcript_biotype}."\t".($info{'exon_number'} || '-')."\t".
       ($tv->{'feature'}{'aa'} || '-')."\t".($tv->{'feature'}{'codons'} || '-')."\t".$strand."\t".($info{transcript_support_level}|| '-');
